Refactor main binary to remove the lib.
