+++ .vscode/settings.json
@@ -67,6 +67,7 @@
         "Uninit",
         "unseparated",
         "vectorially",
+        "VERGEN",
         "WHITESPACES"
     ],
     "cSpell.language": "en-US",
