+++ crates/rsonpath/Cargo.toml
@@ -21,10 +21,6 @@ bench = false
 path = "./src/main.rs"
 name = "rq"
 
-[lib]
-bench = false
-name = "rqlib"
-
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
