+++ crates/rsonpath/src/error.rs
@@ -1,5 +1,3 @@
-#![forbid(unsafe_code)]
-
 use color_eyre::{Help, SectionExt};
 use eyre::eyre;
 use rsonpath_lib::{
