+++ crates/rsonpath/src/logger.rs
@@ -1,0 +1,10 @@
+use color_eyre::eyre::{Result, WrapErr};
+use log::LevelFilter;
+use simple_logger::SimpleLogger;
+
+pub fn configure(verbose: bool) -> Result<()> {
+    SimpleLogger::new()
+        .with_level(if verbose { LevelFilter::Trace } else { LevelFilter::Warn })
+        .init()
+        .wrap_err("Logger configuration error.")
+}
