+++ rsonpath.code-workspace
@@ -60,6 +60,7 @@
 			"Uninit",
 			"unseparated",
 			"vectorially",
+			"VERGEN",
 			"WHITESPACES"
 		],
 		"cSpell.language": "en-US",
